[IPR-36] refactor(company): 도메인 분리하여 경로 이동
