+++ src/main/java/com/flex/interpre/domain/company/controller/CompanyController.java
@@ -1,0 +1,4 @@
+package com.flex.interpre.domain.company.controller;
+
+public class CompanyController {
+}
