+++ src/main/java/com/flex/interpre/domain/company/dto/CompanyDetailResponse.java
@@ -1,0 +1,4 @@
+package com.flex.interpre.domain.company.dto;
+
+public record CompanyResponse() {
+}
