+++ src/main/java/com/flex/interpre/domain/company/service/CompanyService.java
@@ -1,0 +1,4 @@
+package com.flex.interpre.domain.company.service;
+
+public class CompanyService {
+}
