+++ src/main/java/com/flex/interpre/domain/recruitment/entity/Recruitment.java
@@ -1,7 +1,7 @@
 package com.flex.interpre.domain.recruitment.entity;
 
 import com.flex.interpre.domain.recruitment.dto.request.RecruitmentCreateUpdateRequest;
-import com.flex.interpre.domain.user.entity.Company;
+import com.flex.interpre.domain.company.entity.Company;
 import com.flex.interpre.global.constant.Area;
 import com.flex.interpre.global.constant.JobSecond;
 import com.flex.interpre.global.constant.JobFirst;
