+++ src/main/java/com/flex/interpre/domain/user/service/UserService.java
@@ -6,7 +6,7 @@ import com.flex.interpre.domain.user.dto.request.UserUpdateRequest;
 import com.flex.interpre.domain.user.dto.response.MyCompanyInfo;
 import com.flex.interpre.domain.user.dto.response.MyJobSeekerInfo;
 import com.flex.interpre.domain.user.dto.response.MyUserDetailInfo;
-import com.flex.interpre.domain.user.entity.Company;
+import com.flex.interpre.domain.company.entity.Company;
 import com.flex.interpre.domain.user.entity.JobSeeker;
 import com.flex.interpre.domain.user.entity.User;
 import com.flex.interpre.domain.user.exception.UserExceptions;
